changed the layout of some ui elements + added comments and souts to a lot of methods + organized the code by creating new methods for functionalities
