--- src/ecoSolutionsShop/Model/Order.java
@@ -1,10 +1,0 @@
-package ecoSolutionsShop.Model;
-
-public class Order {
-
-    String orderID, orderStatus;
-
-    public  void createPriceSum(){
-
-    }
-}
