+++ src/ecoSolutionsShop/Model/Status.java
@@ -3,11 +3,6 @@ package ecoSolutionsShop.Model;
 public class Status {
 
     public final static String dirtyInShop = "Dirty in shop";
-    public final static String arrivedInCenter = "Arrived in center";
-    public final static String inWashing = "In washing";
-    public final static String inDrying = "In drying";
-    public final static String inIroning = "In ironing";
-    public final static String readyForTransport = "Ready for transport";
     public final static String cleanInShop = "Clean in shop";
     public final static String completed = "Completed";
