+++ src/ecoSolutionsShop/Services/Notification.java
@@ -1,0 +1,5 @@
+package ecoSolutionsShop.Services;
+
+public class Notification {
+
+}
