--- src/ecoSolutionsShop/Services/Notificitaion.java
@@ -1,8 +1,0 @@
-package ecoSolutionsShop.Services;
-
-public class Notificitaion {
-
-    public void sendNotification(){
-
-    }
-}
