+++ src/ecoSolutionsShop/Services/invoices/0.txt
@@ -1,1 +1,1 @@
-Total: 0 kr    Client name:     Shop name:     Client email:     
+Total: 0 kr    Client name: null    Shop name:     Client email: null
