+++ src/ecoSolutionsShop/View/FXML/checkOrder.fxml
@@ -11,7 +11,7 @@
 
 <AnchorPane fx:id="pane" prefHeight="1080.0" prefWidth="1920.0" stylesheets="@../CSS/style.css" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ecoSolutionsShop.Controller.CheckOrderController">
     <children>
-        <ChoiceBox fx:id="orderStatus_choiceBox" focusTraversable="false" layoutX="1478.0" layoutY="496.0" prefHeight="40.0" prefWidth="370.0" />
+        <ChoiceBox fx:id="orderStatus_choiceBox" focusTraversable="false" layoutX="1041.0" layoutY="675.0" prefHeight="40.0" prefWidth="370.0" />
         <TableView fx:id="order_tableview" focusTraversable="false" layoutX="104.0" layoutY="495.0" prefHeight="563.0" prefWidth="760.0" stylesheets="@../CSS/tableView.css">
             <columns>
                 <TableColumn fx:id="description_column" prefWidth="190.0" text="Description" />
@@ -88,7 +88,7 @@
                 <Font size="30.0" />
             </font>
         </Label>
-      <Button fx:id="set_btn" layoutX="1748.0" layoutY="1014.0" mnemonicParsing="false" onAction="#setStatuses" prefHeight="40.0" prefWidth="100.0" stylesheets="@../CSS/button.css" text="Set">
+      <Button fx:id="set_btn" layoutX="1041.0" layoutY="785.0" mnemonicParsing="false" onAction="#setStatuses" prefHeight="40.0" prefWidth="100.0" stylesheets="@../CSS/button.css" text="Set">
          <font>
             <Font size="20.0" />
          </font>
@@ -98,7 +98,7 @@
             <Font size="20.0" />
          </font>
       </Label>
-      <Label layoutX="1478.0" layoutY="457.0" text="Set order status">
+      <Label layoutX="1041.0" layoutY="630.0" text="Set order status">
          <font>
             <Font size="20.0" />
          </font>
